requirements.txt
pybit
requests
python-telegram-bot==13.15
python-binance